display corner even with toolbars and stuff
Many CSS frameworks use z-index extensively for toolbars
and the like. 1000 seems to be the "normal" z-index for this stuff
thus hiding the profiler corner. Unless you know
what you are looking for this is very hard to debug.

This patch sets a unreasonable high z-index
(10001) in the hope to be rendered above everything else.