Sane defaults, no `config.py`
This provides a more aggresive refactoring of configuration than
https://github.com/mdornseif/gae_mini_profiler/tree/appengine_config

Per default everyone on the development server get's profiled and
on production people with access to the AppEngine Admin Console get
profiled.

This can be overwritten by adding a list of E-Mail addresses in
`appengine_config.py`.

This change is insofar incompatible as it removes `config.py`.